App commit (version 1.1)
- Implementation of Mixed Group: from now the parliament members are
fixed and the members who are not assigned (they became not assigned
after removing a party or at the beginnig of the game) are moved to
mixed group and are not controlled by any player.

The interface starts with only one big Mixed Group type party. The user
shall insert one by one the parties that he needs.

The Mixed Group does not vote at the moment.

- Code reorganization and skimming