Test Commitment
Crash test added to project:
- testIllegalAssignment function tests the event for which the user
tries to assign to a party more members than those that are available in
Mixed Group:
FIRST CASE: user tries to assign all members in the parliament;
SECOND CASE: user tries to assign all the members it could receive from

Mixed Group plus one more).
As result of this operation the party will gain all the remaining
members in the Mixed Group but not more.